requirements.txt
pyyaml==6.0.1
pathlib==1.0.1 
boto3==1.34.69 